+++ CHANGELOG.md
@@ -6,9 +6,16 @@ and this project does NOT adhere to [Semantic Versioning](https://semver.org/spe
 
 
 ## [Unreleased]
+### Added
+- Show users any 5xx server error response and extracted failure reason
+- 5xx server errors have a reload button which will initiate a new download with the same URL
+- Adding feed: Cmd+R will reload the same URL
+- Settings, Feeds: Cmd+R will reload the data source 
+
 ### Fixed
 - Changed error message text when user cancels creation of new feed item
 - Comparing existing articles with nonexistent guid and link
+- Adding feed: If URLs can't be resolved in the first run (5xx error), try a second time. E.g., 'Done' click (issue: #5)  
 
 ### Changed
 - Interface builder files replaced with code equivalent